입력, 수정, 삭제 구현 (드래그 기능 작업해야함)

+++ src/main/java/com/clipclap/rego/model/dto/TouristAttractionDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class TouristAttractionDTO {
 
-    private Long touristAttractionId;
+    private Integer touristAttractionId;
 
     private String address;
